removed superflous classes and grouped functions in already existing ones + the same for UI structure files (FXML)
